refactor: Updated filestructure to folder/index
