+++ src/components/LandingPage/Sections/FeaturesSection/index.tsx
