+++ src/components/LandingPage/Sections/GetStartedSection/index.tsx
