+++ src/components/LandingPage/Sections/PartnerSection/index.tsx
