+++ src/components/index.ts
@@ -1,13 +1,13 @@
-import Link from './Link/link';
-import Button from './Button/button';
-import Header from './Header/header';
+import Link from './Link';
+import Button from './Button';
+import Header from './Header';
 
-import CommunitySection from './LandingPage/Sections/CommunitySection/community-section';
-import FeaturesSection from './LandingPage/Sections/FeaturesSection/features-section';
-import GetStartedSection from './LandingPage/Sections/GetStartedSection/getstarted-section';
-import MainSection from './LandingPage/Sections/MainSection/main-section';
-import PartnerSection from './LandingPage/Sections/PartnerSection/partner-section';
-import SandboxSection from './LandingPage/Sections/SandboxSection/sandbox-section';
+import CommunitySection from './LandingPage/Sections/CommunitySection';
+import FeaturesSection from './LandingPage/Sections/FeaturesSection';
+import GetStartedSection from './LandingPage/Sections/GetStartedSection';
+import MainSection from './LandingPage/Sections/MainSection';
+import PartnerSection from './LandingPage/Sections/PartnerSection';
+import SandboxSection from './LandingPage/Sections/SandboxSection';
 
 export {
   Link,
