+++ src/pages/community.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Layout from '../templates/Layout/layout';
+import Layout from '../templates/Layout';
 
 export default function DownloadPage(): JSX.Element {
   const title = 'Download Node.js';
