+++ src/templates/Layout/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Header from '../../components/Header/header';
+import Header from '../../components/Header';
 import SEO from '../../components/seo';
 
 import styles from './layout.module.scss';
