+++ src/templates/learn.tsx
@@ -1,7 +1,7 @@
 import { graphql } from 'gatsby';
 import React from 'react';
 import Article from '../components/article';
-import Layout from './Layout/layout';
+import Layout from './Layout';
 import Navigation from '../components/navigation';
 import { LearnPageContext, LearnPageData } from '../types';
